Refactor up gatsby-node.js.
