+++ gatsby/gatsby-node/onCreateNode.ts
@@ -1,0 +1,15 @@
+import { createFilePath } from "gatsby-source-filesystem"
+
+export const onCreateNode = ({ node, actions, getNode }) => {
+  const { createNodeField } = actions
+
+  if (node.internal.type === `MarkdownRemark`) {
+    const value = createFilePath({ node, getNode })
+
+    createNodeField({
+      name: `slug`,
+      node,
+      value,
+    })
+  }
+}
